% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_long_to_wide}
\alias{convert_long_to_wide}
\title{generate expression matrix with multiple samples}
\usage{
convert_long_to_wide(data, mainCode_, subCode_)
}
\arguments{
\item{data}{fileterd data from spark database}

\item{mainCode_}{mainCode selection}

\item{subCode_}{subCode selection}
}
\value{
a expression matrix with gene row and sample column
}
\description{
convert long to wide for SQL dataframe from spark database
}
