% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{create_mae_airway}
\alias{create_mae_airway}
\title{Create a MultiAssayExperiment object from airway data using DESeq2 analysis.}
\usage{
create_mae_airway()
}
\value{
A MultiAssayExperiment object.
}
\description{
This function loads the airway dataset, performs DESeq2 analysis,
and creates a MultiAssayExperiment (MAE) object. The DESeq2 results
(p-value and log2FoldChange) are attached to the rowData of the RNAseq assay.
}
