% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{create_mae_airway_edgeR}
\alias{create_mae_airway_edgeR}
\title{Create a MultiAssayExperiment object from airway data using edgeR analysis.}
\usage{
create_mae_airway_edgeR()
}
\value{
A MultiAssayExperiment object.
}
\description{
This function loads the airway dataset, performs edgeR differential expression analysis,
and creates a MultiAssayExperiment (MAE) object. The edgeR results (including p-value,
log2FoldChange, and FDR) are attached to the rowData of the RNAseq assay.
}
