% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplotFuns.R
\name{createPCAPlot}
\alias{createPCAPlot}
\title{PCAplot}
\usage{
createPCAPlot(pcaResult, colData, pcX, pcY)
}
\arguments{
\item{pcaResult}{pca table}

\item{colData}{gene symbol}

\item{pcX}{x axis PC dimension}

\item{pcY}{y axis PC dimension}
}
\value{
PCA plot
}
\description{
PCA plot for RNAseq expression
}
