% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spark.R
\name{dbBrowserServer}
\alias{dbBrowserServer}
\title{dbBrowerSever}
\usage{
dbBrowserServer(id, sc)
}
\arguments{
\item{id}{module id for UI and Server}

\item{sc}{spark connection}
}
\value{
database browser Server module
}
\description{
search spark database
}
