% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spark.R
\name{dbBrowserUI}
\alias{dbBrowserUI}
\title{dbBrowerUI}
\usage{
dbBrowserUI(id)
}
\arguments{
\item{id}{module id for UI and Server}
}
\value{
database browser UI module
}
\description{
search spark database
}
