% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqShinyApp.R
\name{demo_oncoExprAppSpark}
\alias{demo_oncoExprAppSpark}
\title{shiny app for RNAseq  demo version}
\usage{
demo_oncoExprAppSpark()
}
\value{
start the UI and Server for analysis
}
\description{
start the shiny app with spark connection for shiny proxy at the seqslab console
}
