% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expr_pattern}
\alias{expr_pattern}
\title{expression pattern analysis}
\usage{
expr_pattern(
  data_v2,
  special_gene_list,
  save_path_ = NULL,
  save_filename_ = "scatter_plot.png",
  cutoff_tpm = 7,
  x_col = "tissue",
  y_col = "ctc"
)
}
\arguments{
\item{data_v2}{dataframe derived from Prep4ExprComp function}

\item{special_gene_list}{gene list derived from Prep4ExprComp function}

\item{save_path_}{where you want to save file}

\item{save_filename_}{plot filename}

\item{cutoff_tpm}{what target genes would be shown in red color}

\item{x_col}{x axis tile of scatter plot}

\item{y_col}{y axis tile of scatter plot}
}
\value{
scatter plot for expression pattern analysis
}
\description{
expression pattern analysis for two samples by scatter plot
}
