% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{filterModuleServer}
\alias{filterModuleServer}
\title{filtering table for analysis (server)}
\usage{
filterModuleServer(id, data)
}
\arguments{
\item{id}{id}

\item{data}{data}
}
\value{
filtered table
}
\description{
filtering table for analysis
}
