% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_selection.R
\name{filterModuleUI}
\alias{filterModuleUI}
\title{filtering table for analysis (UI)}
\usage{
filterModuleUI(id)
}
\arguments{
\item{id}{id}
}
\value{
filtered table
}
\description{
filtering table for analysis
}
