% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_colData}
\alias{generate_colData}
\title{generate colData for RNAseqShinyApp}
\usage{
generate_colData(normcount = normCount, grouplist, genecol = "GeneSymbol")
}
\arguments{
\item{normcount}{normCount table to get colnames (sample id)}

\item{grouplist}{group list}

\item{genecol}{to remove gene column}
}
\value{
colData
}
\description{
generate classification for colData
}
