% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_selection.R
\name{generate_coldata_for_local_tcga_data}
\alias{generate_coldata_for_local_tcga_data}
\title{colData for grouping to analysis}
\usage{
generate_coldata_for_local_tcga_data(filtered_table, mainCode_, subCode_)
}
\arguments{
\item{filtered_table}{table with selecting column}

\item{mainCode_}{main column for patient ID}

\item{subCode_}{sub column for grouping or classification}
}
\value{
filtered table
}
\description{
colData for grouping to analysis
}
