% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_colData_random}
\alias{generate_colData_random}
\title{generate pseudo colData with random classification}
\usage{
generate_colData_random(normcount = normCount, genecol = "gene_id")
}
\arguments{
\item{normcount}{normCount table to get colnames (sample id)}

\item{genecol}{to remove gene column}
}
\value{
pseudo colData
}
\description{
generate random classification for colData
}
