% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkData.R
\name{generate_sample_info}
\alias{generate_sample_info}
\title{sample information table}
\usage{
generate_sample_info(data, mainCode_, subCode_)
}
\arguments{
\item{data}{fileterd data from spark database}

\item{mainCode_}{mainCode selection}

\item{subCode_}{subCode selection}
}
\value{
a table about sample codes
}
\description{
generate sample information table for MAE colData
}
