% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_latest_file_group_df}
\alias{get_latest_file_group_df}
\title{time filter for filename selection}
\usage{
get_latest_file_group_df(filename_list)
}
\arguments{
\item{filename_list}{list of filenames}
}
\value{
filename list with time information
}
\description{
extract the latest file from filename list with time information
}
