% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{ggvolcano_custom}
\alias{ggvolcano_custom}
\title{ggvolcano_custom}
\usage{
ggvolcano_custom(
  df,
  geneName,
  pValCol = "PValue",
  logFCCol = "logFC",
  coef = 2,
  lfc_cut = 1,
  pval_cut = 0.05,
  useAdjP = FALSE,
  title = "Volcano Plot",
  topN = 20,
  geneCol = NULL,
  pointSize = 2,
  ptAlpha = 0.6,
  labelSize = 3
)
}
\arguments{
\item{df}{DEG table}

\item{geneName}{gene symbol}

\item{pValCol}{which col for p-value}

\item{logFCCol}{which col for log fold-change}

\item{coef}{coefficient for limma}

\item{lfc_cut}{log fold-change cut-off label}

\item{pval_cut}{p-value for cut-off label}

\item{useAdjP}{adjusted p-value}

\item{title}{plot title}

\item{topN}{show a number of genes symbol with high log fold-change}

\item{geneCol}{gene symbol column}

\item{pointSize}{point plot size}

\item{ptAlpha}{transparent}

\item{labelSize}{gene symbol text size}
}
\value{
ggplot object
}
\description{
ggvolcano_custom for RNAseq analysis with custom parameters
}
