% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{ggvolcano_limma}
\alias{ggvolcano_limma}
\title{limma analysis with MAE input}
\usage{
ggvolcano_limma(
  fit,
  coef = 2,
  lfc_cut = 1,
  pval_cut = 0.05,
  useAdjP = FALSE,
  title = "Volcano Plot",
  topN = 0,
  geneCol = NULL,
  pointSize = 2,
  ptAlpha = 0.6,
  labelSize = 3,
  ...
)
}
\arguments{
\item{fit}{limma fit result}

\item{coef}{coefficient for limma}

\item{lfc_cut}{log fold-change cut-off label}

\item{pval_cut}{p-value for cut-off label}

\item{useAdjP}{adjusted p-value}

\item{topN}{show a number of genes symbol with high log fold-change}

\item{pointSize}{point plot size}

\item{ptAlpha}{transparent}

\item{labelSize}{gene symbol text size}

\item{assayName}{assay}

\item{subCodeCol}{which col for comparasion}
}
\value{
limma analysis table
}
\description{
differential expression by limma
}
