% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{go_enrich_dotplot}
\alias{go_enrich_dotplot}
\title{Gene Ontology Enrichment Analysis}
\usage{
go_enrich_dotplot(
  gene_list_,
  save_path_ = NULL,
  save_filename_ = NULL,
  mode_ = "BP",
  showCategory_ = 10
)
}
\arguments{
\item{gene_list_}{a set of genes with specific expression level}

\item{save_path_}{what path you want to save the files}

\item{save_filename_}{what name you call the files}

\item{mode_}{Gene Ontology mode such as MF, BP, CC}

\item{showCategory_}{the number of GO term you want to show in doplot, default: 10}
}
\value{
the dotplot for GO enrichment
}
\description{
a set of genes for enrichment analysis with Gene Ontology database
}
