% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_enrichment.R
\name{gseaFCModuleServer}
\alias{gseaFCModuleServer}
\title{gseaeFCModuleServer}
\usage{
gseaFCModuleServer(id, DEG_table, direction = c("up", "down"))
}
\arguments{
\item{id}{module id for UI and Server}

\item{DEG_table}{DEG table}

\item{direction}{direction of GSEA (up or down)}
}
\value{
GSEA module Server
}
\description{
GSEA module
}
