% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{gtf2expr}
\alias{gtf2expr}
\title{gtf to expr profile}
\usage{
gtf2expr(gtf_data)
}
\arguments{
\item{gtf_data}{gtf_data的絕對路徑}
}
\value{
經過處理後的表現量表格，包括Cov, FPKM, TPM三種量化指標
}
\description{
輸入RNA-seq pipline生成的 .gtf file
}
