% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\docType{package}
\name{HeatmapMAE}
\alias{oncoexpr}
\alias{oncoexpr-package}
\alias{HeatmapMAE}
\title{HeatmapMAE Package: Interactive Heatmap from MAE Object}
\description{
This package provides functions to generate interactive heatmaps based on a
MultiAssayExperiment (MAE) object. Two different differential expression analysis
workflows are supported: DESeq2 and edgeR. The resulting MAE object is assumed to contain
the necessary differential expression results (p-value and log2FoldChange) in the rowData
of the RNAseq assay.
}
\section{Example Usage}{

The following examples demonstrate how to use the functions.
\subsection{Using DESeq2 workflow:}{

\dontrun{
  library(HeatmapMAE)
  # Create MAE object using DESeq2 analysis
  mae <- create_mae_airway()
  # Generate heatmap for selected genes (e.g., first 50 genes)
  ht <- make_heatmap_mae(mae, geneList = rownames(mae)[1:50])
  # Run interactive Shiny App
  run_heatmap_app(mae, geneList = rownames(mae)[1:50])
}
}

\subsection{Using edgeR workflow:}{

\dontrun{
  library(HeatmapMAE)
  # Create MAE object using edgeR analysis
  mae_edgeR <- create_mae_airway_edgeR()
  # Generate heatmap for selected genes (e.g., first 50 genes)
  ht_edgeR <- make_heatmap_mae(mae_edgeR, geneList = rownames(mae_edgeR)[1:50])
  # Run interactive Shiny App
  run_heatmap_app(mae_edgeR, geneList = rownames(mae_edgeR)[1:50])
}
}
}

\author{
\strong{Maintainer}: charles chuang \email{charles.chuang@atgenomix.com} (\href{https://orcid.org/0000-0002-4571-3608}{ORCID})

}
