% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneFraction.R
\name{ImmuneFractionPlot}
\alias{ImmuneFractionPlot}
\title{Immune Cell Fraction Prediction}
\usage{
ImmuneFractionPlot(
  group1_,
  group2_,
  groups_list,
  method_ = "quantiseq",
  tumorStatus = TRUE,
  save_path = NULL
)
}
\arguments{
\item{group1_}{sample 1}

\item{group2_}{sample 2}

\item{groups_list}{sample types}

\item{tumorStatus}{if your sample is tumor TRUE or FALSE}

\item{save_path}{what the path you want to save the files}

\item{method}{quantiseq, xCell, CIBERSORT}
}
\value{
plot for immune cell faction of your samples
}
\description{
predict the composition of immune cell types for your samples
}
