% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuneFraction.R
\name{ImmuneFractionPlot_mae}
\alias{ImmuneFractionPlot_mae}
\title{Immune Cell Fraction Prediction}
\usage{
ImmuneFractionPlot_mae(
  mae,
  groups_list,
  method_ = "quantiseq",
  tumorStatus = TRUE,
  save_path = NULL
)
}
\arguments{
\item{mae}{multiple assays experiments}

\item{groups_list}{sample types}

\item{tumorStatus}{if your sample is tumor TRUE or FALSE}

\item{save_path}{what the path you want to save the files}

\item{method}{quantiseq, xCell, CIBERSORT}
}
\value{
plot for immune cell faction of your samples
}
\description{
predict the composition of immune cell types for your samples
}
