% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{input_and_prep_file}
\alias{input_and_prep_file}
\title{prepare data}
\usage{
input_and_prep_file(entirepath, groups_list)
}
\arguments{
\item{entirepath}{entire path for gtf files}

\item{groups_list}{the group name of samples}
}
\value{
expression profile
}
\description{
precessing the gtf file into expression profile
}
