% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{interactivePlotsUI}
\alias{interactivePlotsUI}
\title{Interactive Plots UI}
\usage{
interactivePlotsUI(id)
}
\arguments{
\item{id}{A character string specifying the module ID.}
}
\value{
A Shiny tag list containing the UI layout for interactive plots.
}
\description{
Creates a user interface component for displaying interactive volcano, scatter, and violin plots.
}
\examples{
\dontrun{
  ui <- fluidPage(
    interactivePlotsUI("myPlotModule")
  )
}

}
