% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LimmaMAE}
\alias{LimmaMAE}
\title{limma analysis with MAE input}
\usage{
LimmaMAE(
  mae,
  assayName = "RNAseq",
  subCodeCol = "subCode",
  coef = 2,
  pval_cut = 0.05,
  lfc_cut = 1,
  useAdjP = FALSE
)
}
\arguments{
\item{mae}{multiple assays experiment}

\item{assayName}{assay}

\item{subCodeCol}{which col for comparasion}

\item{coef}{coefficient for limma}

\item{pval_cut}{p-value for cut-off label}

\item{lfc_cut}{log fold-change cut-off label}

\item{useAdjP}{adjusted p-value}
}
\value{
limma analysis table
}
\description{
differential expression by limma
}
