% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{make_heatmap_mae}
\alias{make_heatmap_mae}
\title{Generate a ComplexHeatmap heatmap based on a MAE object with separate RNAseq and DEG assays.}
\usage{
make_heatmap_mae(mae, geneList = NULL)
}
\arguments{
\item{mae}{A MultiAssayExperiment object containing at least two experiments:
one for RNAseq expression (with an assay "normCount") and one for DEG results.}

\item{geneList}{A character vector of gene identifiers to include. If NULL, all genes are used.}
}
\value{
A \code{ComplexHeatmap} object.
}
\description{
This function extracts normalized counts from the RNAseq assay and differential expression
results (p-value and log2FoldChange) from the DEG assay of a MAE object. It subsets the data
based on a provided gene list (if given) and generates a heatmap with three tracks:
the z-score of expression, -log10(p-value), and log2FoldChange.
}
\details{
It assumes that the RNAseq SummarizedExperiment has an assay named "normCount" and that
the DEG SummarizedExperiment contains the differential expression results in its rowData.
}
