% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{make_heatmap_mae_pheatmap}
\alias{make_heatmap_mae_pheatmap}
\title{Generate pheatmap from a MAE Object}
\usage{
make_heatmap_mae_pheatmap(mae, geneList = NULL)
}
\arguments{
\item{mae}{A MultiAssayExperiment object that must contain an experiment named "RNAseq" with a "normCount" assay.}

\item{geneList}{Optional character vector specifying the genes to subset. If NULL, all genes are used.}
}
\value{
A pheatmap object.
}
\description{
This function extracts RNA-seq normalized counts (assay "normCount") from the RNAseq experiment
in a MAE object, optionally subsets the data based on a provided gene list, and retrieves sample
annotation from the colData (defaulting to the "group" column). It then uses the pheatmap package
to generate a heatmap that displays both gene names and sample names.
}
