% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{mod_filter_server}
\alias{mod_filter_server}
\title{filtering table for viewing data (server)}
\usage{
mod_filter_server(id, long_df)
}
\arguments{
\item{id}{id}

\item{long_df}{data with long format}
}
\value{
filtered table
}
\description{
filtering data with long format
}
