% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{mod_filter_sidebar_ui}
\alias{mod_filter_sidebar_ui}
\title{filtering table for viewing data (UI)}
\usage{
mod_filter_sidebar_ui(id)
}
\arguments{
\item{id}{id}
}
\value{
filtered table
}
\description{
filtering table for viewing data
}
