% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sample_selection.R
\name{mod_geneSelector_server}
\alias{mod_geneSelector_server}
\title{DE geneSelector server}
\usage{
mod_geneSelector_server(id, deg_table, geneList)
}
\arguments{
\item{id}{module id for UI and Server}

\item{deg_table}{original DEG table}

\item{geneList}{DEG list}
}
\value{
click gene to render on the volcano plot
}
\description{
click gene from DEG table for plot
}
