% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModule.R
\name{mod_geneSelector_ui}
\alias{mod_geneSelector_ui}
\title{DE geneSelector UI}
\usage{
mod_geneSelector_ui(id)
}
\arguments{
\item{id}{module id for UI and Server}
}
\value{
click gene to render on the volcano plot
}
\description{
click gene from DEG table for plot
}
