% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{mTissueTyper}
\alias{mTissueTyper}
\title{multiple samples TissueTyper}
\usage{
mTissueTyper(
  selectMode = FALSE,
  geneList_ = NULL,
  sample_list_,
  save_path,
  gtexFilePath,
  groups_list = c("tissue", "ctc")
)
}
\arguments{
\item{selectMode}{select or excluding}

\item{geneList_}{target gene set}

\item{sample_list_}{multiple sample id}

\item{save_path}{what the path you want to save the files}

\item{gtexFilePath}{"GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_median_tpm.gct"}

\item{groups_list}{sample types}
}
\value{
heatmap plot for correlation coefficient between multiple tissues of GTEx and your samples
}
\description{
predict the original tissue what the sample from
}
