% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{oncoExprApp}
\alias{oncoExprApp}
\title{shiny app for this analysis}
\usage{
oncoExprApp(host_ = NULL, port_ = NULL)
}
\arguments{
\item{host_}{host name}

\item{port_}{port number}
}
\value{
start the UI and Server for analysis
}
\description{
start the shiny app
}
