% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqShinyApp.R
\name{oncoExprAppSpark}
\alias{oncoExprAppSpark}
\title{shiny app for RNAseq}
\usage{
oncoExprAppSpark(host_ = NULL, port_ = NULL)
}
\arguments{
\item{host_}{host name}

\item{port_}{port number}
}
\value{
start the UI and Server for analysis
}
\description{
start the shiny app with spark connection
}
