% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pca.R
\name{pcaModuleServer}
\alias{pcaModuleServer}
\title{PCA plot Server}
\usage{
pcaModuleServer(id, normCount, colData)
}
\arguments{
\item{id}{module id for UI and Server}

\item{normCount}{gene expression matrix}

\item{colData}{sample information}
}
\value{
PCA plot Server
}
\description{
PCA
}
