% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{Prep4ExprComp}
\alias{Prep4ExprComp}
\title{prepare for expression pattern}
\usage{
Prep4ExprComp(group1, group2, groups_name = c("tissue", "ctc"))
}
\arguments{
\item{group1}{first sample}

\item{group2}{sencond sample}

\item{groups_name}{sample type name e.g. tissue & ctc}
}
\value{
a list with dataframe and specific gene list
}
\description{
compare two samples' expression pattern
}
