% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_progress_popup.R
\name{progressPopupServer}
\alias{progressPopupServer}
\title{progressPopupServer}
\usage{
progressPopupServer(id)
}
\arguments{
\item{id}{module id for UI and Server}
}
\value{
progress popup Server
}
\description{
progress popup Server
}
