% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModule.R
\name{progressPopupUI}
\alias{progressPopupUI}
\title{progressPopupUI}
\usage{
progressPopupUI(id)
}
\arguments{
\item{id}{module id for UI and Server}
}
\value{
progress popup UI
}
\description{
progress popup UI
}
