% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{RNAseqShinyAppSpark}
\alias{RNAseqShinyAppSpark}
\title{shiny app for RNAseq for public use}
\usage{
RNAseqShinyAppSpark()
}
\value{
start the UI and Server for analysis
}
\description{
start the RNAseq shiny app with spark connection for shiny proxy at the seqslab console
}
