% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{run_heatmap_app}
\alias{run_heatmap_app}
\title{Run a Shiny App to display the interactive heatmap.}
\usage{
run_heatmap_app(mae, geneList = NULL)
}
\arguments{
\item{mae}{A MultiAssayExperiment object.}

\item{geneList}{A character vector of gene identifiers to include in the heatmap. If NULL, all genes will be used.}
}
\description{
This function launches a Shiny app that displays the heatmap generated from a MAE object
using a specified gene list. The UI is minimal and only shows the heatmap output.
}
