% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModule.R
\name{sampleSelectionServer}
\alias{sampleSelectionServer}
\title{sampleSelect shinmy Server module}
\usage{
sampleSelectionServer(id, sc, selected_db, selected_table, set_colData)
}
\arguments{
\item{id}{module id for UI and Server}

\item{sc}{spark connection}

\item{selected_db}{select database from spark}

\item{selected_table}{select deltatable from selected database}

\item{set_colData}{a reactive Value for storage dynamic dataframe}
}
\value{
Sample selection Server module
}
\description{
set shiny Server module for sample selection
}
