% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyModule.R
\name{sampleSelectionUI}
\alias{sampleSelectionUI}
\title{sampleSelect shinmy UI module}
\usage{
sampleSelectionUI(id)
}
\arguments{
\item{id}{module id for UI and Server}
}
\value{
UI module
}
\description{
set shiny UI module for sample selection
}
