% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkData.R
\name{se2longformat}
\alias{se2longformat}
\title{se to longformat}
\usage{
se2longformat(
  se_ = exp_brca,
  assay_name = "fpkm_unstrand",
  selected_row_ = c(1:5),
  selected_col_ = c("TCGA-5L-AAT0-01A-12R-A41B-07", "TCGA-A2-A04U-01A-11R-A115-07"),
  selected_colData_ = "tissue_type"
)
}
\arguments{
\item{se_}{se object}

\item{assay_name}{selecting assay from se object}

\item{selected_row_}{select gene}

\item{selected_col_}{select sample}

\item{selected_colData_}{filter colData column}
}
\value{
pivot long format from se object
}
\description{
se from MAE package transfered to long format dataframe
}
