% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkData.R
\name{show_spark_tables}
\alias{show_spark_tables}
\title{view spark}
\usage{
show_spark_tables(sc, use_dataset)
}
\arguments{
\item{sc}{spark connection}

\item{use_dataset}{selecting dataset to show tables}
}
\value{
tables in the datasets
}
\description{
show spark datasets and tables
}
