% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spark.R
\name{sparkConnectionServer}
\alias{sparkConnectionServer}
\title{Server for connecting to spark}
\usage{
sparkConnectionServer(id)
}
\arguments{
\item{id}{id}
}
\value{
pivot long format from se object
}
\description{
spark connection shiny server
}
