% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_spark.R
\name{sparkConnectionUI}
\alias{sparkConnectionUI}
\title{UI for connecting to spark}
\usage{
sparkConnectionUI(id)
}
\arguments{
\item{id}{id}
}
\value{
pivot long format from se object
}
\description{
spark connection shiny UI
}
