% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{target_exprofile}
\alias{target_exprofile}
\title{target gene expression}
\usage{
target_exprofile(
  selectMode = TRUE,
  geneList_,
  groups_list_,
  save_path_ = NULL,
  save_filename_ = "target_gene_expr.csv",
  expr_profile_
)
}
\arguments{
\item{selectMode}{select or exclude to gene list}

\item{geneList_}{interesting genes}

\item{groups_list_}{term of group}

\item{save_path_}{save path}

\item{save_filename_}{save name}

\item{expr_profile_}{target gene expression profile}
}
\value{
the expression of genes which you are interesting
}
\description{
針對目標基因回傳兩個樣本的基因表現量
}
