% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoexpr.R
\name{TissueTyper}
\alias{TissueTyper}
\title{TissueTyper}
\usage{
TissueTyper(
  selectMode = FALSE,
  geneList_ = NULL,
  group1_,
  group2_,
  save_path,
  gtexFilePath,
  groups_list = c("tissue", "ctc"),
  sample_id = "BC021-RNA"
)
}
\arguments{
\item{group1_}{sample 1}

\item{group2_}{sample 2}

\item{save_path}{what the path you want to save the files}

\item{gtexFilePath}{"GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_median_tpm.gct"}

\item{groups_list}{sample types}

\item{sample_id}{samply id}
}
\value{
heatmap plot for correlation coefficient between multiple tissues of GTEx and your samples
}
\description{
predict the original tissue what the sample from
}
