% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplotFuns.R
\name{createPCAPlot}
\alias{createPCAPlot}
\title{PCAplot}
\usage{
createPCAPlot(pcaResult, colData, enableClustering = FALSE, centers = 2)

createPCAPlot(pcaResult, colData, enableClustering = FALSE, centers = 2)
}
\arguments{
\item{pcaResult}{pca table}

\item{colData}{gene symbol}

\item{enableClustering}{set clustering or not}

\item{centers}{centers number (default = 2)}

\item{pcX}{x axis PC dimension}

\item{pcY}{y axis PC dimension}
}
\value{
PCA plot

PCA plot
}
\description{
PCA plot for RNAseq expression

PCA plot for RNAseq expression
}
