% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqShinyApp.R
\name{RNAseqShinyAppSpark}
\alias{RNAseqShinyAppSpark}
\title{shiny app for RNAseq for public use}
\usage{
RNAseqShinyAppSpark()
}
\value{
A Shiny application object.
}
\description{
start the RNAseq shiny app with spark connection for shiny proxy at the seqslab console
}
\examples{
\dontrun{
  library(sparklyr)
  sc <- spark_connect(master = "local")
  input_data <- read.csv("path/to/rnaseq_data.csv")
  RNAseqShinyAppSpark(input_data, sc, "RNAseq Analysis", 8080)
}

}
